Refactored TCPClient class into multiple sub-classes #11 #13 #14 #16 #18 #19 #20
